+++ cedar-policy/CHANGELOG.md
@@ -14,6 +14,7 @@ Starting with version 3.2.4, changes marked with a star (*) are _language breaki
 
 ### Added
 - Public syntax tree (`pst`) support for `variadic-is-in-range` feature: a variadic `isInRange` is modelled by a `pst::Expr::VariadicOp{...}` in the PST (#2380).
+- Functions for inspecting the determining policies on a TPE Response (`TpeResponse::decision`, `TpeResponse::satisfied_permits`, and `TpeResponse::satisfied_forbids`)
 
 ### Changed